Fix grid layout
Previously, "display: grid" was only used for the actual buttons.
That led to inconsistencies between the size of the buttons and
the display on viewport change.

Now, the app uses "display: grid" on both the buttons and the
display with the display taking the entire first row. This
is accomplished using "grid-column: 1 / -1" on the display.