Debug operator buttons
Previously, if there was a previous operand and the user pressed an
operator twice without any current operand, a NaN error would pop up.

Now, the app introduces an if statement to handle this case and what's
more, it now allows for changing the operator. For example, if a user
wanted to add 2 and 2 together but pressed a different operator, they
can now easily switch to the plus operator.